Seed script, market calendar, market data simulator, and a bug-class test
npm run seed -- --reset stands up believable demo data from zero: 4 model
portfolios (versioned), 5 instruments, 314 deterministic daily closes with one
deliberately withheld, 4 customers (2 approved / 1 pending / 1 rejected),
deposits settling T+1, notional buys allocating model weights by
largest-remainder, dividends across ex-date and pay-date, and a FIFO sell
spanning two lots. Ledger balances in all six commodities.

The June VXUS lot is fully exhausted by that sell and gives up exactly its
$4,998.00 opening basis -- the no-drift rule holding on real data rather than
in a unit test.

Market calendar computes NYSE holidays from the rules rather than a list that
expires: 4 July 2026 is a Saturday and correctly observes on Friday 3 July, and
a Friday 4 Sep trade settles Tuesday 8 Sep across Labor Day. 17 tests.

Market data downgraded live -> simulated and labelled everywhere: Alpaca Broker
sandbox keys are not entitled to the data API (401 on Basic, 401 on APCA
headers, 404 on the broker proxy). The brief permits this slot to be simulated;
the three mandatory live slots are unaffected. Owning the price source is also
what makes a corrected close issuable on demand.

sql-hygiene.test.ts ends a bug class rather than its instances: sum() over a
bigint column returns numeric, which our parser leaves as a string. I wrote that
bug three times before writing the test, which then found two more. It was
nearly invisible -- the first instance reported 'trial balance nets to zero'
only because the table was empty and [].every() is true.

Also: pg type parsers moved out of db.ts into pg-types.ts imported explicitly,
because a script importing Pool from pg directly got the default parsers; and
kyc ordering now breaks ties on id, because now() is the transaction timestamp
so rows written together share it exactly.

+++ src/lib/ledger/invariants.ts
@@ -12,6 +12,7 @@
  * entire point. It is exposed both as `npm run verify` and as a page.
  */
 
+import '../pg-types';
 import { Client } from 'pg';
 
 export interface InvariantCheck {
@@ -276,7 +277,7 @@ export async function runInvariants(connectionString?: string): Promise<Invarian
 
     const { rows: totals } = await client.query<{
       commodity: string; cents: string | null; units: string | null;
-    }>(`SELECT commodity, sum(amount_cents) AS cents, sum(units) AS units
+    }>(`SELECT commodity, sum(amount_cents)::bigint AS cents, sum(units) AS units
           FROM journal_lines GROUP BY commodity ORDER BY commodity`);
 
     for (const row of totals) {
